requirement.txt
pytest>=7.4.0
pyyaml
requests>=2.31.0
loguru>=0.7.0